Sales Order page - Move vue script to resources assets component
